Plant + Poster Check (#184)
## About The Pull Request
This is a PR that adds plants and posters to the Extera_Colony map and
more specifically adds a wealth of plants or posters in just about every
area that I felt it would make some sense to do so in order to aid with
sanity gain and as requested by Sigmasquad. Moreover, I added a pair of
new rooms; small room tweaks where I thought it prudent; two named
plants; and a small little map feature (On top of the bar) and spruced
up a pre-existing one (Sandro hard-hat one) since it didn't make sense
with the roof being a thing now but I added a hole.

## Changelog
:cl: Aurke
add: Added a pair of new maint rooms that keep in the plant or poster
theme to sow confusion.
add: Added some new plants/posters to every Z-level in a fair amount of
rooms for sanity gain and slight realism.
tweak: Deleted some existed posters or moved around some things in order
to keep from bloating the number while adding a lot.
add: Created two departmental plants by the names of Jeremy and Dakota
for a bit of roleplay fluff. **These should be fine as I believe I
remembered to do quotes but please note I am new and could have fucked
something up.** _SHOULD_ be fine though.
tweak: Miscellaneous edits to things like errant walls or other such
things I felt made more sense that are too small to remember.
/:cl:

That should detail basically everything, I hope that this is good for a
first PR and thank you for reading!